requirements.txt
ultralytics~=8.3.27
flask~=3.0.3
requests~=2.32.3
jsonify
pymongo~=4.10.1
opencv-python~=4.10.0.84
numpy~=1.24.3
scipy~=1.9.1
pillow~=9.3.0
deepface~=0.0.80
tensorflow~=2.12.0
cloudinary~=1.41.0
python-dotenv
pytest~=8.3.3
beautifulsoup4~=4.12.3
soupsieve~=2.6
dnspython~=2.7.0
idna~=3.7
httpcore~=1.0.7
httpx~=0.27.2
cryptography~=43.0.3
anyio~=4.6.2.post1
jax~=0.4.30
jaxlib~=0.4.30
flatbuffers~=24.3.25
pygments~=2.18.0
filelock~=3.16.1
colorama~=0.4.6
matplotlib~=3.9.2
defusedxml~=0.7.1
cffi~=1.17.1
pip~=24.3.1
attrs~=24.2.0
wheel~=0.41.2
jinja2~=3.1.4
rsa~=4.9
pyasn1~=0.6.1
fire~=0.7.0
termcolor~=2.5.0
gast~=0.4.0
h5py~=3.12.1
pytz~=2024.2
torch~=2.5.1
rich~=13.0.1
keras~=2.12.0
PyYAML~=6.0.2
zipp~=3.20.2
MarkupSafe~=3.0.2
werkzeug~=3.1.1
click~=8.1.7
blinker~=1.8.2
itsdangerous~=2.2.0
tqdm~=4.66.6
six~=1.16.0
pandas~=2.2.3
wrapt~=1.14.1
mtcnn~=0.1.1
Cython~=3.0.11
setuptools~=75.5.0
mpmath~=1.3.0
sympy~=1.13.1
inference~=0.27.0
pydot~=2.0.0
astunparse~=1.6.3
safetensors~=0.4.5
onnxruntime~=1.19.2
tensorboard~=2.12.3
packaging~=24.1
torchvision~=0.20.1
psutil~=6.1.0
urllib3~=1.26.20
fsspec~=2024.10.0
aiohttp~=3.10.11
yarl~=1.17.2
cachetools~=5.5.0
certifi~=2024.8.30
python-dateutil~=2.9.0.post0
pywin32~=308
py-cpuinfo~=9.0.0
seaborn~=0.13.2
cycler~=0.12.1
gdown~=5.2.0
gunicorn~=23.0.0
Markdown~=3.7
networkx~=3.2.1
oauthlib~=3.2.2
contourpy~=1.3.0
fonttools~=4.54.1
pyparsing~=3.2.0
kiwisolver~=1.4.7
botocore~=1.35.63
boto3~=1.35.60
av~=12.3.0
shapely~=2.0.6